feat: add loginSubject

+++ youtube-client-app/src/app/auth/services/auth.service.ts
@@ -1,39 +1,44 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
+import { BehaviorSubject } from 'rxjs';
 import { SearchDataService } from 'src/app/youtube/services/search-data.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthService {
+  loginSubject = new BehaviorSubject<boolean>(false)
+
   constructor(
     private router: Router,
     private searchDataService: SearchDataService,
-  ) { }
+  )
+  {
+    if (localStorage.getItem('login')) {
+      this.loginSubject.next(true)
+    }
+    this.loginSubject.subscribe({
+      next: (val) => {
+        if (val) {
+          localStorage.setItem('login', 'test');
+          this.router.navigate(['']);
+        } else {
+          localStorage.removeItem('login');
+          this.router.navigate(['login']);
+          this.searchDataService.deleteResultData();
+        }
+      },
+      error: (err) => console.log({"err": err})
+    })
+  }
 
   public login(): void {
-    localStorage.setItem('login', 'test');
-    this.router.navigate(['']);
+    this.loginSubject.next(true)
   }
 
   public logout(): void {
-    localStorage.removeItem('login');
-    this.router.navigate(['login']);
-    this.searchDataService.deleteResultData();
+    this.loginSubject.next(false)    
   }
-
-  // static isAuthCheck(): Promise<boolean> {
-  //   return new Promise((resolve) => {
-  //     setTimeout(() => {
-  //       const isAuth = localStorage.getItem('login');
-  //       if (isAuth === 'test') {
-  //         resolve(true);
-  //       } else {
-  //         resolve(false);
-  //       }
-  //     }, 0);
-  //   });
-  // }
 
   static isAuthCheck(): boolean {
     const isAuth = localStorage.getItem('login');
